io and print work. lab 2 compiles. added test driver script for io. changed u32 to unsigned long, which solved the type problems. int is 16 bits using bcc compiler.

+++ my_types.h
@@ -1,9 +1,8 @@
 #ifndef MY_TYPES_H
 #define MY_TYPES_H
 
-// these are correct for my mac and Mint 64-bit Lisa 12. YMMV
 typedef unsigned char  u8;
 typedef unsigned short u16;
-typedef unsigned int   u32;
+typedef unsigned long   u32;
 
 #endif
